apps/api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
psycopg2-binary==2.9.7 
sqlalchemy==2.0.21
asyncpg==0.29.0
alembic==1.12.1
pydantic>=2.0.0,<3.0.0 
fastapi-users[sqlalchemy,jwt]==12.1.3
passlib[bcrypt]==1.7.4 
fastapi-users-db-sqlalchemy==7.0.0 
boto3==1.34.48 
httpx>=0.28.1,<1.0.0
google-genai==1.19.0